Imported run into the Gui
I renamed the gui file to GUI.py. It was FloatTest.py before. I also moved the GUI.py and its .kv file (my.kv) to the root of the directory. The pictures are still in their own directory but we can now import other files without problems. The pictures are easily accessed as well using GUIPics/namofpic.png.  I commented everything out in run.py except for the declaration and call of the main function as well as a print to screen so I could tell run.py was running. When I tried to run run.py without commenting everything else I got errors about finding modules associated with the robot arms.